// apps/main/src/models/v2/controlplane/environment.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { EnvironmentStatus } from './environmentStatus';

/**
 * An environment.
 */
export interface Environment {
  /** Environment identifier */
  id: string;
  /** Unique uid for the environment to identify a unique lifecycle */
  uuid: string;
  /** Project identifier */
  project_id: string;
  /** Unique uid for the project to identify a unique lifecycle */
  project_uuid?: string;
  /** The environment type for the environment. */
  env_type_id: string;
  /** Environment human readable name */
  display_name: string;
  /** The date and time when the resource type was created */
  created_at: string;
  /** The date and time when the environment was updated. */
  updated_at: string;
  status: EnvironmentStatus;
  /** An optional message associated with the status. */
  status_message?: string;
  /** The id of the runner to be used to deploy this environment. If not set, this will be filled in upon first deployment. */
  runner_id?: string;
}